requirement.txt
Django>=5.0.6,<5.1
djangorestframework>=3.15.1,<3.16
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
djangorestframework-simplejwt
django-environ
